Move form into dedicated component
Also validate form on select option
