+++ clients/admin-ui/src/features/custom-fields/CustomFieldModal.tsx
@@ -1,39 +1,23 @@
 import {
-  Box,
-  Button,
-  Flex,
-  IconButton,
   Modal,
   ModalBody,
   ModalContent,
   ModalHeader,
   ModalOverlay,
-  SimpleGrid,
-  Text,
 } from "@fidesui/react";
-import {
-  FIELD_TYPE_OPTIONS,
-  RESOURCE_TYPE_OPTIONS,
-  FIELD_TYPE_OPTIONS_NEW,
-  FieldTypes,
-} from "common/custom-fields";
-import CustomInput, {
-  CUSTOM_LABEL_STYLES,
-} from "common/custom-fields/form/CustomInput";
-import { AddIcon } from "common/custom-fields/icons/AddIcon";
-import FormSection from "common/form/FormSection";
-import { CustomSelect } from "common/form/inputs";
+import { FieldTypes } from "common/custom-fields";
 import { getErrorMessage } from "common/helpers";
 import { useAlert } from "common/hooks";
-import { TrashCanSolidIcon } from "common/Icon/TrashCanSolidIcon";
-import { FieldArray, Form, Formik, FormikHelpers } from "formik";
+import { Formik } from "formik";
+import { useState } from "react";
 import * as Yup from "yup";
 
+import { CustomFieldForm } from "~/features/custom-fields/CustomFieldForm";
 import {
   useAddCustomFieldDefinitionMutation,
-  useUpsertAllowListMutation,
   useGetAllowListQuery,
   useUpdateCustomFieldDefinitionMutation,
+  useUpsertAllowListMutation,
 } from "~/features/plus/plus.slice";
 import {
   AllowedTypes,
@@ -42,12 +26,6 @@ import {
   CustomFieldDefinitionWithId,
   ResourceTypes,
 } from "~/types/api";
-import { useMemo, useState } from "react";
-
-const CustomFieldLabelStyles = {
-  ...CUSTOM_LABEL_STYLES,
-  minWidth: "unset",
-};
 
 type ModalProps = {
   isOpen: boolean;
@@ -56,7 +34,7 @@ type ModalProps = {
   customField?: CustomFieldDefinitionWithId;
 };
 
-type FormValues = Omit<CustomFieldDefinition, "field_type"> & {
+export type FormValues = Omit<CustomFieldDefinition, "field_type"> & {
   allow_list: AllowListUpdate;
   field_type: FieldTypes;
 };
@@ -113,7 +91,7 @@ export const CustomFieldModal = ({
       skip: !customField?.allow_list_id,
     });
 
-  let [validationSchema, setNewValidationSchema] = useState(
+  const [validationSchema, setNewValidationSchema] = useState(
     optionalValidationSchema
   );
 
@@ -130,7 +108,7 @@ export const CustomFieldModal = ({
       } as unknown as FormValues)
     : initialValuesTemplate;
 
-  const handelDropdownChange = (value: string) => {
+  const handelDropdownChange = (value: FieldTypes) => {
     if (value === FieldTypes.OPEN_TEXT) {
       setNewValidationSchema(optionalValidationSchema);
     } else {
@@ -138,9 +116,11 @@ export const CustomFieldModal = ({
     }
   };
 
+  // const test = useFormikContext<FormValues>()
+  // test.
   const handleSubmit = async (
-    values: FormValues,
-    helpers: FormikHelpers<FormValues>
+    values: FormValues
+    // helpers: FormikHelpers<FormValues>
   ) => {
     if (
       [FieldTypes.SINGLE_SELECT, FieldTypes.MULTIPLE_SELECT].includes(
@@ -177,7 +157,7 @@ export const CustomFieldModal = ({
         // eslint-disable-next-line no-param-reassign
         values.allow_list_id = result.data?.id;
       }
-      //Add error handling here
+      // Add error handling here
     }
 
     if (values.field_type === FieldTypes.OPEN_TEXT) {
@@ -214,7 +194,7 @@ export const CustomFieldModal = ({
       successAlert(`Custom field successfully saved`);
     }
   };
-  console.log("initialValues", initialValues);
+  // console.log("initialValues", initialValues);
   return (
     <Modal
       id="custom-field-modal-hello-world"
@@ -255,165 +235,17 @@ export const CustomFieldModal = ({
             initialValues={initialValues}
             validationSchema={validationSchema}
             onSubmit={handleSubmit}
+            enableReinitialize
+            validateOnChange
           >
-            {({
-              dirty,
-              isValid,
-              isSubmitting,
-              values,
-              errors,
-              validateForm,
-            }) => (
-              <Form
-                style={{
-                  paddingTop: "12px",
-                  paddingBottom: "12px",
-                }}
-              >
-                <Box py={3}>
-                  <FormSection title="Field Information">
-                    <CustomInput
-                      displayHelpIcon
-                      isRequired
-                      label="Name"
-                      name="name"
-                      customLabelProps={CustomFieldLabelStyles}
-                    />
-                    <CustomInput
-                      displayHelpIcon
-                      label="Description"
-                      name="description"
-                      customLabelProps={CustomFieldLabelStyles}
-                    />
-                    <CustomSelect
-                      label="Location"
-                      name="resource_type"
-                      options={RESOURCE_TYPE_OPTIONS}
-                      labelProps={CustomFieldLabelStyles}
-                    />
-                  </FormSection>
-                </Box>
-                <Box py={3}>
-                  <FormSection title="Configuration">
-                    <CustomSelect
-                      label="Field Type"
-                      name="field_type"
-                      options={FIELD_TYPE_OPTIONS_NEW}
-                      onChange={async (e) => {
-                        handelDropdownChange(e.value);
-                        await validateForm();
-                      }}
-                    />
-                    {values.field_type !== FieldTypes.OPEN_TEXT ? (
-                      <Flex
-                        flexDirection="column"
-                        gap="12px"
-                        paddingTop="6px"
-                        paddingBottom="24px"
-                      >
-                        <FieldArray
-                          name="allow_list.allowed_values"
-                          render={(fieldArrayProps) => {
-                            // eslint-disable-next-line @typescript-eslint/naming-convention
-                            const { allowed_values } = values.allow_list;
-                            // @ts-ignore
-                            return (
-                              <Flex flexDirection="column" gap="24px" pl="24px">
-                                <Flex flexDirection="column">
-                                  {allowed_values.map((_value, index) => (
-                                    <Flex
-                                      flexGrow={1}
-                                      gap="12px"
-                                      // eslint-disable-next-line react/no-array-index-key
-                                      key={index}
-                                      mt={index > 0 ? "12px" : undefined}
-                                      // ref={fieldRef}
-                                    >
-                                      <CustomInput
-                                        customLabelProps={{
-                                          color: "gray.600",
-                                          fontSize: "sm",
-                                          fontWeight: "500",
-                                          lineHeight: "20px",
-                                          minW: "126px",
-                                          pr: "8px",
-                                        }}
-                                        displayHelpIcon={false}
-                                        isRequired
-                                        label={`List item ${index + 1}`}
-                                        name={`allow_list.allowed_values[${index}]`}
-                                      />
-                                      <IconButton
-                                        aria-label="Remove this list value"
-                                        data-testid={`remove-list-value-btn-${index}`}
-                                        icon={<TrashCanSolidIcon />}
-                                        isDisabled={allowed_values.length <= 1}
-                                        onClick={() =>
-                                          fieldArrayProps.remove(index)
-                                        }
-                                        size="sm"
-                                        variant="ghost"
-                                      />
-                                    </Flex>
-                                  ))}
-                                </Flex>
-                                <Flex alignItems="center">
-                                  <Text
-                                    color="gray.600"
-                                    fontSize="xs"
-                                    fontWeight="500"
-                                    lineHeight="16px"
-                                    pr="8px"
-                                  >
-                                    Add a list value
-                                  </Text>
-                                  <IconButton
-                                    aria-label="Add a list value"
-                                    data-testid="add-list-value-btn"
-                                    icon={<AddIcon h="7px" w="7px" />}
-                                    onClick={() => {
-                                      fieldArrayProps.push("");
-                                    }}
-                                    size="xs"
-                                    variant="outline"
-                                  />
-                                  {allowed_values.length === 0 &&
-                                  errors?.allow_list ? (
-                                    <Text color="red.500" pl="18px" size="sm">
-                                      {errors.allow_list.allowed_values}
-                                    </Text>
-                                  ) : null}
-                                </Flex>
-                              </Flex>
-                            );
-                          }}
-                        />
-                      </Flex>
-                    ) : null}
-                  </FormSection>
-                </Box>
-
-                <SimpleGrid columns={2} width="100%">
-                  <Button
-                    variant="outline"
-                    mr={3}
-                    onClick={onClose}
-                    data-testid="cancel-btn"
-                    isDisabled={isLoading || isSubmitting}
-                  >
-                    Cancel
-                  </Button>
-                  <Button
-                    type="submit"
-                    colorScheme="primary"
-                    data-testid="save-btn"
-                    isLoading={isLoading}
-                    isDisabled={!dirty || !isValid || isSubmitting}
-                  >
-                    Save
-                  </Button>
-                </SimpleGrid>
-              </Form>
+            {(props) => (
+              <CustomFieldForm
+                validationSchema={validationSchema}
+                isLoading={isLoading}
+                onClose={onClose}
+                handleDropdownChange={handelDropdownChange}
+                {...props}
+              />
             )}
           </Formik>
         </ModalBody>
