+++ clients/admin-ui/src/features/custom-fields/cells.tsx
@@ -22,8 +22,8 @@ import {
 import { useUpdateCustomFieldDefinitionMutation } from "~/features/plus/plus.slice";
 import {
   CustomFieldDefinition,
-  ScopeRegistryEnum,
   ResourceTypes,
+  ScopeRegistryEnum,
 } from "~/types/api";
 
 export const ResourceTypeCell = (
